Fix the constant reskipping after the initial skip
It happened because of a questionable behavior of dataclasses, which
creates a __hash__() method implicitly.

However, the issue of sessions before resimulated until the next regular
notfication is still happening, and is only guarded by
AutoSkipper.accept_session(). So in the end, the visible behavior is as
intended, but still we shouldn't be emitting those simulated
notifications (because we skipped). To deal with later.