rlz script: correctly rollback pyproject commit
Previously the rollback of the pyproject.toml git commit would only
restore the HEAD^ version of the file. This would NOT undo the commit
itself. As a further consequence of this, the subsequent
'git checkout HEAD -- pyproject.toml' to rollback 'poetry version' would
restore the comitted pyproject.toml (which commit did not get undone).